+++ CHANGELOG.md
@@ -45,6 +45,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - [#458](https://github.com/babylonlabs-io/babylon/pull/458) Set `CosmosProvider` functions as public in
 `babylonclient` package
 
+### State Machine Breaking
+
+- [#519](https://github.com/babylonlabs-io/babylon/pull/519) Add missing data in `InitGenesis` and `ExportGenesis` in `x/incentive` module
+
 ### Bug fixes
 
 - [#509](https://github.com/babylonlabs-io/babylon/pull/509) crypto: fix ECDSA malleability